Reset inline styles on animateBack
Removing `transform` from inline styles is preferrable to setting to 'none', as this is less prescriptive to the consumer.

`transition` is being removed rather set to `10ms` as this will presist on the element after you drag it again. In other words, the first time you drag a card, there will be no transistion on the transform, but after animateBack is trigger, transistion 10ms will presist, which in incurs performance overhead and makes the dragging process laggy.